Code refactoring. Moved "Extensions" to single project as primary modules. Simplified including DbContext logic.
