+++ Pendulum/Pendulum.CQS/Builders/CommandBuilder.cs
@@ -1,8 +1,5 @@
-using System;
 using System.Threading.Tasks;
-using Microsoft.Extensions.DependencyInjection;
 using Pendulum.CQS.Contexts;
-using Pendulum.CQS.Extensions.ServiceProvider;
 using Pendulum.CQS.Interfaces;
 
 namespace Pendulum.CQS.Builders
