+++ Pendulum/Pendulum.CQS/Interfaces/IPendulumContainer.cs
@@ -1,4 +1,7 @@
+using System;
+using System.Collections.Generic;
 using System.Reflection;
+using Microsoft.Extensions.DependencyInjection;
 
 namespace Pendulum.CQS.Interfaces
 {
